optimize: lift off at feed rate
to help protect the drill bit when drilling.  This is a non-optimal solution, since most times while milling we could lift-off at move speed, but I don't want to take time right now to sort out the code enough for that, so I'm just having it use feed speed every time, and relying on it not to take too long, since it's not very far and not that often.

This includes a couple of updates to make this not break visualize.py